Create Bracket match judgment
This problem is used to find out whether the parentheses in the expression match. Just determine whether the parentheses in the expression (there are only three parentheses that will appear in this question, which are parentheses, parentheses, and curly braces) match, and there can be other values in the expression or not.